pyproject.toml
[tool.poetry]
name = "cyberolympusapp"
version = "0.1.0"
description = "Site"
authors = ["Pavel14701 <pavelkutia@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
authlib = "1.3.2"
flask = "^3.0.3"
flask-babel = "^4.0.0"
flask-sqlalchemy = "3.1.1"
flask-login = "^0.6.3"
flask-compress = "^1.15"
flask-wtf = "1.2.1"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
sqlalchemy = "2.0.34"
redis = "5.0.8"
turbo-flask = "0.8.5"
flask-caching = "^2.3.0"
flask-session = "^0.8.0"
cryptography = "^43.0.1"
charset-normalizer = "3.3.2"
babel = "^2.16.0"
blinker = "^1.8.2"
brotli = "^1.1.0"
cachelib = "0.9.0"
certifi = "2024.8.30"
cffi = "1.17.1"
click = "8.1.7"
colorama = "0.4.6"
flask-cdn = "^1.5.3"
flask-sock = "0.7.0"
greenlet = "3.0.3"
h11 = "0.14.0"
idna = "3.8"
itsdangerous = "2.2.0"
jinja2 = "3.1.4"
markupsafe = "2.1.5"
msgspec = "0.18.6"
oauthlib = "2.1.0"
pycparser = "2.22"
pytz = "2024.2"
requests = "2.32.3"
requests-oauthlib = "1.1.0"
simple-websocket = "1.0.0"
speaklater = "1.3"
typing-extensions = "4.12.2"
urllib3 = "2.2.2"
werkzeug = "3.0.4"
wsproto = "1.2.0"
wtforms = "3.1.2"
zstandard = "0.23.0"
libsass = "^0.23.0"
flask-assets = "^2.1.0"
webassets = "^2.0"
watchdog = "^5.0.2"
packaging = "^24.1"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
